Maintainance version (#28)
* 5 add mixed order method (#23)

* add first attempt - no tests

* fix returned order

* add test for first and second

* add liteature regression test

* add method to docs

* fix spelling

* add roy 2003 ref. standardize citation style

* add docstring to test

* PR feedback on f_0

* remove union operator fix imports

* split discretization into manageable file sizes

* consistent capitalization
